coned-scraper/python-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pyotp>=2.9.0
playwright>=1.48.0
python-multipart>=0.0.12
pydantic>=2.10.0
cryptography>=43.0.0
aiohttp>=3.11.0
paho-mqtt>=2.1.0

